GeneSets

GeneSets.get_msigdb — Method

get_msigdb(version="v7.2")

This function automatically fetches msigdb gene sets from the Broad Institute and generate an array of GeneSet objects containing relevant information Input: version: The version of the database, starts with "v" (e.g. "v.7.2"). Currently defaulted to Version.

GeneSets.get_msigdb

GeneSets.get_msigdb

GeneSets.get_msigdb — Function

get_msigdb(version="v7.2")

This function automatically fetches msigdb gene sets from the Broad Institute and generate an array of GeneSet objects containing relevant information Input: version: The version of the database, starts with "v" (e.g. "v.7.2"). Currently defaulted to Version.